PMM-13250 move restore test to own service
